requirements.txt
# Contract Analysis & Risk Assessment Bot
# Dependencies for GenAI-powered Legal Assistant

# Core Framework
streamlit>=1.31.0

# LLM Integration
openai>=1.12.0

# NLP Processing
spacy>=3.7.0
nltk>=3.8.1

# Document Processing
PyPDF2>=3.0.1
python-docx>=1.1.0
pdfplumber>=0.10.3

# PDF Generation
fpdf2>=2.7.8

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Multilingual Support
langdetect==1.0.9
# Note: googletrans removed due to httpx conflict with openai
# Using built-in language detection in src/nlp/language_detector.py instead

# Visualization
plotly>=5.18.0

# Utilities
python-dateutil>=2.8.2
python-dotenv>=1.0.0
